% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_models.R
\name{rescaled_est_4a}
\alias{rescaled_est_4a}
\title{Rescale linear estimates when sampling scheme satisfies assumption 4a}
\usage{
rescaled_est_4a(y, x, z, H, db, nb)
}
\arguments{
\item{y}{Dependent variable in the regression.}

\item{x}{Vector of shocks.}

\item{z}{Matrix of covariates. If you want to include an intercept, you need to add column of ones here.}

\item{H}{Adjacency matrix of sampled network.}

\item{db}{Mean missing degree for nodes with at least one incorrectly sampled link.}

\item{nb}{Number of nodes with at least one incorrectly sampled link.}
}
\value{
Vector of coefficient estimates, and vector of standard errors (computed using
normal lm procedure) assuming no uncertainty in estimate of db. To get coefficients
assuming some uncertainty in db, you need to use the bootstrap function.
}
\description{
This function computes bias-corrected OLS estimates of spillover effects from
sampled network data under the assumption that the distribution of sampled
degrees and unsampled degrees are independent for node with at least one
incorrectly sampled link. In cases where they are dependent, see rescaled_est_4b.
}
