% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_models.R
\name{rescaled_est_4b}
\alias{rescaled_est_4b}
\title{Rescale linear estimates when sampling scheme satisfies assumption 4b}
\usage{
rescaled_est_4b(y, x, H, db, dh, nb)
}
\arguments{
\item{y}{Dependent variable in the regression.}

\item{x}{Vector of shocks.}

\item{H}{Adjacency matrix of sampled network.}

\item{db}{Vector of missing degree for each node with at least one incorrectly sampled link.}

\item{dh}{Vector of sampled degree for each node with at least one incorrectly sampled link}

\item{z}{Matrix of covariates. If you want to include an intercept, you need to add column of ones here.}
}
\value{
Vector of coefficient estimates, and vector of standard errors (computed using
normal lm procedure) assuming no uncertainty in estimate of db. To get coefficients
assuming some uncertainty in db, you need to use the bootstrap function.
}
\description{
This function computes bias-corrected OLS estimates of spillover effects from
sampled network data under the assumption that the distribution of sampled
degrees and unsampled degrees are not independent for node with at least one
incorrectly sampled link. In cases where they are independent, see rescaled_est_4a.
}
