% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear_models.R
\name{rescaled_est_5}
\alias{rescaled_est_5}
\title{Rescale nonlinear estimators under assumption 5}
\usage{
rescaled_est_5(y, x, z, H, db, nb)
}
\arguments{
\item{y}{Dependent variable in the regression.}

\item{x}{Vector of shocks.}

\item{z}{Matrix of covariates. If you want to include an intercept, you need to add column of ones here.}

\item{H}{Adjacency matrix of sampled network.}

\item{db}{Mean missing degree for nodes with at least one incorrectly sampled link.}
}
\value{
Vector of coefficient estimates, and vector of standard errors (computed using
normal lm procedure) assuming no uncertainty in estimate of db. To get coefficients
assuming some uncertainty in db, you need to use the bootstrap function.
}
\description{
This function computes bias-corrected estimates of spillover effects from non-linear models with
sampled network data under the assumption that the distribution of sampled
and unsampled links are independent. As instruments, use neighbours of up to degree 2.
}
