% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstraps.R
\name{uniform_boostrap_un}
\alias{uniform_boostrap_un}
\title{Bootstrap estimators for uncertainty of bias-corrected estimators for spillover
effects.}
\usage{
uniform_boostrap_un(y, x, z, H, Mask_mat, num_missing, nos)
}
\arguments{
\item{y}{Dependent variable in the regression.}

\item{x}{Vector of shocks.}

\item{z}{Matrix of covariates. If you want to include an intercept, you need to add column of ones here.}

\item{H}{Adjacency matrix of sampled network.}

\item{nos}{Number of bootstrap estimates.}

\item{mask_mat}{Binary matrix of positions of possibly incorrectly sampled links.}
}
\value{
Standard error of estimates.
}
\description{
This function computes bootstrap estimates of uncertainty for bias-corrected
estimators of spillover effects for linear regression estimators when network
is undersampled.
}
